ADD AreAllPropertiesSet() in CustomerEntity class
In order to properly perform all functions on the main page, a procedure is needed to check whether all information has been properly retrieved in the customer table.
This method checks with a foreach statement whether all this properties are filled in properly and returns false if an empty property is found.