Create EmailValidationPage and Generating an OTP Code for validation
If the email customer enters the ForgetPasswordPage, the customer will be moved to the EmailValidationPage.
In EmailValidationPage, the customer needs to enter a randomly generated OTP code for their email validation.
However due to the limits of developing environment, no way to send a real email to the customer, so just assuming the user receives the OTP code or not.